+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+
+`v1.15.0`_ (2018-12-21)
+~~~~~~~~~~~~~~~~~~~~~~~
+
 - Actually made use of the timeout specified as
   ``FEINCMS_THUMBNAIL_CACHE_TIMEOUT`` instead of the hardcoded value of
   seven days.
@@ -14,6 +18,10 @@ Change log
   updated inlines.
 - Fixed ``TranslatedObjectManager.only_language`` to evaluate callables
   before filtering.
+- Changed the ``render`` protocol of content types to allow returning a
+  tuple of ``(ct_template, ct_context)`` which works the same way as
+  `feincms3's template renderers
+  <https://feincms3.readthedocs.io/en/latest/guides/rendering.html>`__.
 
 
 `v1.14.0`_ (2018-08-16)
@@ -48,4 +56,5 @@ Change log
 
 
 .. _v1.14.0: https://github.com/feincms/feincms/compare/v1.13.0...v1.14.0
-.. _Next version: https://github.com/feincms/feincms/compare/v1.14.0...master
+.. _v1.15.0: https://github.com/feincms/feincms/compare/v1.14.0...v1.15.0
+.. _Next version: https://github.com/feincms/feincms/compare/v1.15.0...master